src/app/core/services/auth/auth-guard.guard.ts
import { Injectable } from '@angular/core';
import { CanActivate, Router } from '@angular/router';
import { AuthService } from './auth.service';
import { RoutePathsConfig } from '../../../app.routes';

@Injectable({
  providedIn: 'root',
})
export class AuthGuard implements CanActivate {
  constructor(private authService: AuthService, private router: Router) {}

  canActivate(): boolean {
    if (this.authService.isLoggedIn()) {
      return true;
    } else {
      // todo: create routePath config with routes
      this.router.navigate([RoutePathsConfig.login]);
      return false;
    }
  }
}
